Expose substring ranges for attributes of custom componets (#26)
This exposes sub-string ranges for attributes of custom components allowing them to locate the exact portion of the source they read from.

Included is an example using this in Dioxus to allow a counter to modify a count stored in the input markdown source.
Making this example work cleanly required some extra utilities for working with sub-strings in Dioxus which are also exported via ReadWriteBox.